+++ example/effects/main.dart
@@ -4,9 +4,9 @@ import 'package:deact/deact_html52.dart';
 void main() {
   deact(
       '#root',
-      globalStateProvider<int>(name: 'render', initialValue: 0, children: [
-        globalStateProvider<int>(name: 'counter', initialValue: 0, children: [
-          globalStateProvider<bool>(name: 'toggle', initialValue: true, children: [
+      globalState<int>(name: 'render', initialValue: 0, children: [
+        globalState<int>(name: 'counter', initialValue: 0, children: [
+          globalState<bool>(name: 'toggle', initialValue: true, children: [
             toggleButton(),
             counterButton(),
             renderButton(),