+++ lib/src/deact/fragment.dart
@@ -1,15 +1,15 @@
 part of deact;
 
-/// A [Fragment] is a [Node] with no content except its
+/// A [FragmentNode] is a [Node] with no content except its
 /// children.
 ///
 /// You can wrap a list of nodes into a fragment and use
 /// the fragment where only a single node is applicable.
-class Fragment extends Node {
-  Fragment._(List<Node> children) : super._(children);
+class FragmentNode extends Node {
+  FragmentNode._(List<Node> children) : super._(children);
 }
 
 /// Creates a [Fragement] node with the given [children].
-Fragment fragment(List<Node> children) {
-  return Fragment._(children);
+FragmentNode fragment(List<Node> children) {
+  return FragmentNode._(children);
 }