+++ lib/src/deact/node.dart
@@ -3,8 +3,8 @@ part of deact;
 /// A Deact application consists of a hierarchy of
 /// [Node]s. A node can have one or more child nodes.
 abstract class Node {
-  final List<Node> children;
+  final List<Node> _children;
   _TreeLocation _location;
 
-  Node._(this.children);
+  Node._(this._children);
 }